refactor: group tests by CLI, adapter, and integration responsibilities
