+++ tests/confluence/test_traversal_incremental.py
